+++ ddSendFeedback_snippet.php
@@ -1,13 +1,13 @@
 <?php
 /**
  * ddSendFeedback
- * @version 2.7.1 (2021-11-09)
+ * @version 2.8 (2024-07-13)
  * 
  * @see README.md
  * 
- * @link https://code.divandesign.biz/modx/ddsendfeedback
+ * @link https://code.divandesign.ru/modx/ddsendfeedback
  * 
- * @copyright 2010–2021 DD Group {@link https://DivanDesign.biz }
+ * @copyright 2010–2024 Ronef {@link https://Ronef.ru }
  */
 
 //Include (MODX)EvolutionCMS.libraries.ddTools